api/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
python-multipart
requests
numpy
pillow
python-jose
passlib
bcrypt==3.2.2
sqlalchemy